Move proxy tests to separate file
