+++ tests/KsuidTest.php
@@ -23,11 +23,6 @@ use DateTimeZone;
 
 class KsuidTest extends TestCase
 {
-    public function testGenerateShouldReturnKsuidInstance()
-    {
-        $this->assertInstanceOf(Ksuid::class, Ksuid::generate());
-    }
-
     public function testShouldBeTrue()
     {
         $this->assertTrue(true);
@@ -49,51 +44,5 @@ class KsuidTest extends TestCase
     {
         $ksuid = new Ksuid;
         $this->assertEquals(27, strlen((string)$ksuid));
-    }
-
-    /* https://segment.com/blog/a-brief-history-of-the-uuid/ */
-    public function testShouldCreateFromString()
-    {
-        $ksuid = KsuidProxy::fromString("0o5Fs0EELR0fUjHjbCnEtdUwQe3");
-        $this->assertEquals("0o5Fs0EELR0fUjHjbCnEtdUwQe3", (string) $ksuid);
-        $this->assertEquals(94985761, $ksuid->timestamp());
-        $this->assertEquals(1494985761, $ksuid->unixtime());
-        $this->assertEquals(
-            "d7b6fe8cd7cff211704d8e7b9421210b",
-            bin2hex($ksuid->payload())
-        );
-
-        $datetime = (new DateTimeImmutable)
-            ->setTimestamp($ksuid->unixtime())
-            ->setTimeZone(new DateTimeZone("UTC"));
-
-        $this->assertEquals(
-            "2017-05-17 01:49:21",
-            $datetime->format("Y-m-d H:i:s")
-        );
-    }
-
-    public function testShouldCreateFromBytes()
-    {
-        $binary = hex2bin("05a95e21d7b6fe8cd7cff211704d8e7b9421210b");
-        $ksuid = KsuidProxy::fromBytes($binary);
-
-        $this->assertEquals("0o5Fs0EELR0fUjHjbCnEtdUwQe3", (string) $ksuid);
-        $this->assertEquals($binary, $ksuid->bytes());
-        $this->assertEquals(94985761, $ksuid->timestamp());
-        $this->assertEquals(1494985761, $ksuid->unixtime());
-        $this->assertEquals(
-            "d7b6fe8cd7cff211704d8e7b9421210b",
-            bin2hex($ksuid->payload())
-        );
-
-        $datetime = (new DateTimeImmutable)
-            ->setTimestamp($ksuid->unixtime())
-            ->setTimeZone(new DateTimeZone("UTC"));
-
-        $this->assertEquals(
-            "2017-05-17 01:49:21",
-            $datetime->format("Y-m-d H:i:s")
-        );
     }
 }
